Reuse CreateBasicFoodDialog
