+++ src/Components/Recipe.js
@@ -16,9 +16,7 @@ import DialogContent from "@mui/material/DialogContent";
 import DialogActions from "@mui/material/DialogActions";
 import Dialog from "@mui/material/Dialog";
 import TextField from "@mui/material/TextField";
-import InputLabel from "@mui/material/InputLabel";
 import MenuItem from "@mui/material/MenuItem";
-import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
 import Chip from "@mui/material/Chip";
 import Autocomplete from "@mui/material/Autocomplete";
@@ -34,29 +32,11 @@ import {
   updateRequest,
   getCalculateFoodSectionForOptions,
   constructBasicFoodOptions,
+  waitForElm,
 } from "../utils";
 
 import { unknownSectionName } from "../constants";
-
-function waitForElm(selector) {
-  return new Promise((resolve) => {
-    if (document.querySelector(selector)) {
-      return resolve(document.querySelector(selector));
-    }
-
-    const observer = new MutationObserver(() => {
-      if (document.querySelector(selector)) {
-        resolve(document.querySelector(selector));
-        observer.disconnect();
-      }
-    });
-
-    observer.observe(document.body, {
-      childList: true,
-      subtree: true,
-    });
-  });
-}
+import CreateBasicFoodDialog from "./CreateBasicFoodDialog";
 
 function Recipe(props) {
   const {
@@ -957,103 +937,6 @@ function Recipe(props) {
     </Dialog>
   );
 
-  const renderCreateBasicFoodDialog = () => (
-    <Dialog
-      open={openCreateBasicFoodDialog}
-      sx={{ "& .MuiDialog-paper": { width: "80%" } }}
-      maxWidth="xs"
-    >
-      <DialogTitle color="primary">Create a new basic food</DialogTitle>
-      <DialogContent dividers>
-        <Stack
-          key={"createBasicFood"}
-          direction="row"
-          justifyContent="space-around"
-          alignItems="center"
-          spacing={2}
-        >
-          <TextField
-            variant="outlined"
-            label={"Food name"}
-            size="small"
-            sx={{ width: "150px" }}
-            value={createBasicFood.name || ""}
-            onChange={(event) => {
-              setCreateBasicFood((previous) => ({
-                ...previous,
-                name: event.target.value,
-              }));
-            }}
-            inputProps={{
-              autoCapitalize: "none",
-            }}
-          />
-          <FormControl size="small" variant="standard">
-            <InputLabel id="tag" style={{ top: "-11px" }}>
-              Dept.
-            </InputLabel>
-            <Select
-              labelId="tag"
-              id="tag"
-              value={createBasicFood.tagId || ""}
-              onChange={(event) => {
-                setCreateBasicFood((previous) => ({
-                  ...previous,
-                  tagId: event.target.value,
-                }));
-              }}
-              style={{ marginTop: 0, paddingTop: "5px", width: "110px" }}
-            >
-              {(glossary && glossary.basicFoodTags
-                ? basicFoodTagOrder.map((basicFoodTagKey) => (
-                    <MenuItem value={basicFoodTagKey} key={basicFoodTagKey}>
-                      {glossary.basicFoodTags[basicFoodTagKey]}
-                    </MenuItem>
-                  ))
-                : []
-              ).concat(
-                <MenuItem value={""} key={"delete"}>
-                  <em>None</em>
-                </MenuItem>
-              )}
-            </Select>
-          </FormControl>
-        </Stack>
-      </DialogContent>
-      <DialogActions>
-        <Button
-          color="secondary"
-          onClick={() => {
-            setOpenCreateBasicFoodDialog(false);
-            setCreateBasicFood({});
-          }}
-        >
-          Cancel
-        </Button>
-        <Button
-          color="success"
-          disabled={!createBasicFood.name}
-          onClick={() => {
-            const foodId = createKey(`${glossaryPath}/basicFoods`);
-            const updates = {};
-            updates[`${glossaryPath}/basicFoods/${foodId}`] =
-              createBasicFood.name;
-            if (createBasicFood.tagId) {
-              updates[`${basicFoodTagAssociationPath}/${foodId}`] =
-                createBasicFood.tagId;
-            }
-            updateRequest(updates);
-            addIngredient(foodId);
-            setOpenCreateBasicFoodDialog(false);
-            setCreateBasicFood({});
-          }}
-        >
-          Create
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-
   return (
     <div>
       <Stack
@@ -1071,7 +954,20 @@ function Recipe(props) {
         </Stack>
       </Stack>
       {renderDeleteDialog()}
-      {renderCreateBasicFoodDialog()}
+      <CreateBasicFoodDialog
+        open={openCreateBasicFoodDialog}
+        createBasicFood={createBasicFood}
+        setCreateBasicFood={setCreateBasicFood}
+        handleSelectedFood={addIngredient}
+        onClose={() => {
+          setOpenCreateBasicFoodDialog(false);
+          setCreateBasicFood({});
+        }}
+        glossary={glossary}
+        basicFoodTagOrder={basicFoodTagOrder}
+        glossaryPath={glossaryPath}
+        basicFoodTagAssociationPath={basicFoodTagAssociationPath}
+      />
     </div>
   );
 }
